refactor: add extension-tag component
